core: Removed some repeated code from particle_data
