+++ src/core/debug.cpp
@@ -158,15 +158,6 @@ void check_particles() {
 
   CELL_TRACE(fprintf(stderr, "%d: entering check_particles\n", this_node));
 
-  /* check the consistency of particle_nodes */
-  /* to this aim the array is broadcasted temporarily */
-  if (this_node != 0)
-    particle_node = (int *)Utils::malloc((max_seen_particle + 1) * sizeof(int));
-  is_here = (int *)Utils::malloc((max_seen_particle + 1) * sizeof(int));
-  memset(is_here, 0, (max_seen_particle + 1) * sizeof(int));
-
-  MPI_Bcast(particle_node, max_seen_particle + 1, MPI_INT, 0, comm_cart);
-
   /* checks: part_id, part_pos, local_particles id */
   for (c = 0; c < local_cells.n; c++) {
     cell = local_cells.cell[c];
@@ -180,8 +171,6 @@ void check_particles() {
             this_node, c, n, cell->part[n].p.identity);
         errexit();
       }
-
-      is_here[part[n].p.identity] = 1;
 
       for (dir = 0; dir < 3; dir++) {
         if (PERIODIC(dir) && (part[n].r.p[dir] < -skin2 ||
@@ -197,12 +186,6 @@ void check_particles() {
                         "id %d: local: %p cell: %p in cell %d\n",
                 this_node, part[n].p.identity,
                 local_particles[part[n].p.identity], &part[n], c);
-        errexit();
-      }
-      if (particle_node[part[n].p.identity] != this_node) {
-        fprintf(stderr,
-                "%d: check_particles: ERROR: node for particle %d wrong\n",
-                this_node, part[n].p.identity);
         errexit();
       }
     }
@@ -235,37 +218,5 @@ void check_particles() {
     errexit();
   }
 
-  /* check whether the particles on my node are actually here */
-  for (p = 0; p <= max_seen_particle; p++) {
-    if (particle_node[p] == this_node) {
-      if (!is_here[p]) {
-        fprintf(stderr, "%d: check_particles: ERROR: particle %d on this node, "
-                        "but not in local cell\n",
-                this_node, p);
-      }
-    }
-  }
-
-  free(is_here);
-
-  if (this_node != 0) {
-    free(particle_node);
-    particle_node = NULL;
-  } else {
-    /* check whether the total count of particles is ok */
-    c = 0;
-    for (p = 0; p <= max_seen_particle; p++)
-      if (particle_node[p] != -1)
-        c++;
-    if (c != n_part) {
-      fprintf(stderr,
-              "%d: check_particles: #particles in particle_node inconsistent\n",
-              this_node);
-      errexit();
-    }
-    CELL_TRACE(fprintf(stderr,
-                       "%d: check_particles: %d particles in particle_node.\n",
-                       this_node, c));
-  }
   CELL_TRACE(fprintf(stderr, "%d: leaving check_particles\n", this_node));
 }
